+++ example/lib/main.dart
@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
 import 'dart:async';
-import 'dart:convert';
 
 import 'package:flutter/services.dart';
 import 'package:native_mixpanel/native_mixpanel.dart';
 
 void main() => runApp(MyApp(
   mixpanel: Mixpanel(
-    isDebug: true,
+    shouldLogEvents: true,
     isOptedOut: false,
   ),
 ));
@@ -40,8 +39,8 @@ class _MyAppState extends State<MyApp> {
     String eventName = 'First App Open';
     // Platform messages may fail, so we use a try/catch PlatformException.
     try {
-      await widget.mixpanel.initialize('<your-token-here>');
-      await widget.mixpanel.track(eventName, jsonEncode({'Math': 'divide'}));
+      await widget.mixpanel.initialize('2500d0d99b3d441c731588f2ef57ef36');
+      await widget.mixpanel.track(eventName, {'Math': 'divide'});
       initStatus = 'Event Sent: $eventName';
     } on PlatformException {
       initStatus = 'Failed to send event: $eventName';
@@ -64,12 +63,85 @@ class _MyAppState extends State<MyApp> {
         appBar: AppBar(
           title: const Text('Plugin example app'),
         ),
-        body: Center(
-          child: Text('$_statusMessage\n'),
+        body: ListView(
+          children: <Widget>[
+            Center(
+              child: Text(
+                '$_statusMessage\n',
+                style: TextStyle(
+                  fontSize: 12.0,
+                  height: 18.0 / 12.0,
+                ),
+              ),
+            ),
+            Container(
+              alignment: Alignment.center,
+              child: RaisedButton(
+                onPressed: () async {
+                  await widget.mixpanel.identify('101');
+                },
+                child: Text('Set Identity'),
+              ),
+            ),
+            Container(
+              alignment: Alignment.center,
+              child: RaisedButton(
+                onPressed: () async {
+                  await widget.mixpanel.alias('premium');
+                },
+                child: Text('Set Alias'),
+              ),
+            ),
+            Container(
+              alignment: Alignment.center,
+              child: RaisedButton(
+                onPressed: () async {
+                  await widget.mixpanel.setPeopleProperties({'genre': 'rock'});
+                },
+                child: Text('Set People Properties'),
+              ),
+            ),
+            Container(
+              alignment: Alignment.center,
+              child: RaisedButton(
+                onPressed: () async {
+                  await widget.mixpanel.registerSuperProperties({'gender': 'he'});
+                },
+                child: Text('Register Super Properties'),
+              ),
+            ),
+            Container(
+              alignment: Alignment.center,
+              child: RaisedButton(
+                onPressed: () async {
+                  await widget.mixpanel.track('TICK', {'createdAt': DateTime.now().toIso8601String()});
+                },
+                child: Text('Send Tick Event'),
+              ),
+            ),
+            Container(
+              alignment: Alignment.center,
+              child: RaisedButton(
+                onPressed: () async {
+                  await widget.mixpanel.reset();
+                },
+                child: Icon(Icons.restore),
+              ),
+            ),
+            Container(
+              alignment: Alignment.center,
+              child: RaisedButton(
+                onPressed: () async {
+                  await widget.mixpanel.flush();
+                },
+                child: Text('Flush'),
+              ),
+            ),
+          ],
         ),
         floatingActionButton: FloatingActionButton(
           onPressed: () {
-            widget.mixpanel.track('Added to Cart', jsonEncode({'ProductId': 'product-${count++}'}));
+            widget.mixpanel.track('Added to Cart', {'ProductId': 'product-${count++}'});
           },
           child: Icon(Icons.plus_one),
         ),