+++ CHANGELOG.md
@@ -16,3 +16,7 @@
 - Fix: props not received in iOS platform.
 
 Force converted arguments to [String:String] type for time being
+
+# 0.0.6
+- Added identify, alias, setPeopleProperties, registerProperties, reset and flush platform calls
+- iOS: Updated reading properties from method arguments